build: modified DELETE persons route to require athorized user to be logged in.

+++ controllers/persons.js
@@ -37,6 +37,20 @@ personsRouter.get('/:id', async (request, response, next) => {
 
 personsRouter.delete('/:id', async (request, response, next) => {
   try {
+    // check and decode user token - it returns Object on which token was based on.
+    // note that token is place in request.token by middleware tokenExtractor
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token invalid' })
+    }
+    // find user trying to delete document
+    const user = await User.findById(decodedToken.id)
+    // check if user is admin
+    if (!user.admin) {
+      return response.status(401).json({ error: 'user unauthorized to delete' })
+    }
+
     const person = await Person.findByIdAndRemove(request.params.id)
 
     response.status(204).end()
